+++ src/commonMain/kotlin/dev/inmo/krontab/utils/SchedulerFlow.kt
@@ -2,9 +2,9 @@ package dev.inmo.krontab.utils
 
 import com.soywiz.klock.DateTime
 import com.soywiz.klock.DateTimeTz
-import dev.inmo.krontab.KronScheduler
-import dev.inmo.krontab.next
-import kotlinx.coroutines.*
+import dev.inmo.krontab.*
+import kotlinx.coroutines.FlowPreview
+import kotlinx.coroutines.delay
 import kotlinx.coroutines.flow.*
 
 /**
@@ -12,30 +12,41 @@ import kotlinx.coroutines.flow.*
  * time zones
  *
  * @see channelFlow
+ * @see KronSchedulerTz.doWhile
  */
 @FlowPreview
-fun KronScheduler.asTzFlow(): Flow<DateTimeTz> = channelFlow {
-    var previousTime = DateTime.nowLocal()
-    while (isActive) {
-        val now = DateTime.nowLocal()
-        val nextTime = next(now) ?: break
-        if (previousTime == nextTime) {
-            delay(1L) // skip 1ms
-            continue
-        } else {
-            previousTime = nextTime
-        }
-        val sleepDelay = (nextTime - DateTime.now().local).millisecondsLong
-        delay(sleepDelay)
-        send(nextTime)
+fun KronSchedulerTz.asTzFlow(): Flow<DateTimeTz> = channelFlow {
+    doInfinity {
+        send(DateTime.nowLocal())
     }
 }
 
 /**
  * This method is a map for [asTzFlow] and will works the same but return flow with [DateTime]s
+ *
+ * @see channelFlow
+ * @see KronScheduler.doWhile
  */
 @FlowPreview
-fun KronScheduler.asFlow(): Flow<DateTime> = asTzFlow().map { it.local }
+fun KronScheduler.asFlow(): Flow<DateTime> = channelFlow {
+    doInfinity {
+        send(DateTime.now())
+    }
+}
+
+/**
+ * This [Flow] will trigger emitting each near time which will be returned from [this] [KronScheduler] with attention to
+ * time zones
+ *
+ * @see channelFlow
+ * @see KronScheduler.asFlow
+ * @see KronSchedulerTz.asTzFlow
+ */
+@FlowPreview
+fun KronScheduler.asTzFlow(): Flow<DateTimeTz> = when (this) {
+    is KronSchedulerTz -> asTzFlow()
+    else -> asFlow().map { it.local }
+}
 
 @Deprecated(
     "It is not recommended to use this class in future. This functionality will be removed soon",